Move waitlist moderation functions into RTK slice
